(refactor) split draftPoints and draftMessages off from points and messages state * added allIds attribute of draftMessage and draftPoints * renamed getPointById to getPointIfReference * added new getPointById and getMessageById utility functions(refactor) split draftPoints and draftMessages off from points and messages state * added allIds attribute of draftPoints * renamed getPointById to getPointIfReference * added new getPointById and getMessageById utility functions * some utility functions now take appState instead of appState.points * replaced isPersisted with isDraft (the opposite) * removed unused isPersisted attribute from components/SemanticScreen.tsx
