+++ src/reducers/points.ts
@@ -17,34 +17,8 @@
   along with U4U.  If not, see <https://www.gnu.org/licenses/>.
 */
 import { Action, Actions } from "../actions/constants";
-import produce from "immer";
-import {
-  PointI,
-  PointReferenceI,
-  isPointShape,
-} from "../dataModels/dataModels";
-import {
-  getPointById,
-  getReferenceData,
-  getReferencedPointId,
-  isReference,
-} from "../dataModels/pointUtils";
+import { PointI, PointReferenceI } from "../dataModels/dataModels";
 import { AppState } from "./store";
-import {
-  _PointCreateParams,
-  PointUpdateParams,
-  _PointsMoveToMessageParams,
-  PointsMoveWithinMessageParams,
-  PointsDeleteParams,
-  CombinePointsParams,
-  _SplitIntoTwoPointsParams,
-} from "../actions/pointsActions";
-import {
-  _MessageCreateParams,
-  //TODO: Does it matter if I pass MessageDeleteParams to the points reducer,
-  //which should never make use of the optional newMessageId prop?
-  MessageDeleteParams,
-} from "../actions/messagesActions";
 
 export interface PointsState {
   byId: {
@@ -59,254 +33,5 @@ export const pointsReducer = (
   action: Action,
   appState: AppState
 ): PointsState => {
-  let newState = state;
-  switch (action.type) {
-    case Actions.pointCreate:
-      newState = handlePointCreate(state, action as Action<_PointCreateParams>);
-      break;
-    case Actions.pointUpdate:
-      newState = handlePointUpdate(state, action as Action<PointUpdateParams>);
-      break;
-    case Actions.pointsMoveWithinMessage:
-      newState = handlePointsMoveWithinMessage(
-        state,
-        action as Action<PointsMoveWithinMessageParams>,
-        appState
-      );
-      break;
-    case Actions.pointsMoveToMessage:
-      newState = handlePointsMoveToMessage(
-        state,
-        action as Action<_PointsMoveToMessageParams>,
-        appState
-      );
-      break;
-    case Actions.pointsDelete:
-      newState = handlePointsDelete(
-        state,
-        action as Action<PointsDeleteParams>,
-        appState
-      );
-      break;
-    case Actions.combinePoints:
-      newState = handleCombinePoints(
-        state,
-        action as Action<CombinePointsParams>,
-        appState
-      );
-      break;
-    case Actions.splitIntoTwoPoints:
-      newState = handleSplitIntoTwoPoints(
-        state,
-        action as Action<_SplitIntoTwoPointsParams>
-      );
-      break;
-    case Actions.messageCreate:
-      newState = handleMessageCreate(
-        state,
-        action as Action<_MessageCreateParams>
-      );
-      break;
-    case Actions.messageDelete:
-      newState = handleMessageDelete(
-        state,
-        action as Action<MessageDeleteParams>,
-        appState
-      );
-      break;
-  }
-  return newState;
+  return state;
 };
-
-function handlePointCreate(
-  state: PointsState,
-  action: Action<_PointCreateParams>
-): PointsState {
-  return produce(state, (draft) => {
-    draft.byId[action.params.newPointId] = {
-      ...action.params.point,
-      _id: action.params.newPointId,
-      pointDate: new Date(),
-    };
-  });
-}
-
-function handlePointUpdate(
-  state: PointsState,
-  action: Action<PointUpdateParams>
-): PointsState {
-  return produce(state, (draft) => {
-    draft.byId[action.params.point._id] = action.params.point;
-  });
-}
-
-function handlePointsMoveWithinMessage(
-  state: PointsState,
-  action: Action<PointsMoveWithinMessageParams>,
-  appState: AppState
-): PointsState {
-  if (appState.drag.context === null) return state;
-  const { region } = appState.drag.context;
-
-  if (!isPointShape(region)) return state;
-  const pointIdsExcludingReferencePoints = appState.selectedPoints.pointIds.filter(
-    (p) => !getReferenceData(p, state)
-  );
-  return produce(state, (draft) => {
-    pointIdsExcludingReferencePoints.forEach(
-      (id) =>
-        (draft.byId[id] = {
-          ...draft.byId[id],
-          shape: region,
-        })
-    );
-  });
-}
-
-function handlePointsMoveToMessage(
-  state: PointsState,
-  action: Action<_PointsMoveToMessageParams>,
-  appState: AppState
-): PointsState {
-  const { newReferencePoints } = action.params;
-
-  if (newReferencePoints === undefined) {
-    return state;
-  }
-
-  return produce(state, (draft) => {
-    newReferencePoints.forEach((point) => {
-      draft.byId[point._id] = point;
-    });
-  });
-}
-
-function handlePointsDelete(
-  state: PointsState,
-  action: Action<PointsDeleteParams>,
-  appState: AppState
-): PointsState {
-  let pointIds = action.params.pointIds;
-  if (action.params.deleteSelectedPoints) {
-    pointIds = pointIds.concat(appState.selectedPoints.pointIds);
-  }
-
-  return produce(state, (draft) => {
-    pointIds.forEach((id) => delete draft.byId[id]);
-  });
-}
-
-//TODO: reuse withinBounds and isQuoted logic in src/reducers/message
-//Is it good form to access appState in src/actions/pointsActions?
-//note: in src/reducers/message, state must be replaced with
-//appState.points (and a similar change in this file)
-function handleCombinePoints(
-  state: PointsState,
-  action: Action<CombinePointsParams>,
-  appState: AppState
-): PointsState {
-  const currentMessage =
-    appState.messages.byId[appState.semanticScreen.currentMessage];
-  const withinBounds = (index: number): boolean => {
-    return (
-      index >= 0 && index < currentMessage.shapes[action.params.shape].length
-    );
-  };
-
-  const isQuoted = (index: number): boolean => {
-    const pointId = currentMessage.shapes[action.params.shape][index];
-    return !!getReferencedPointId(pointId, state);
-  };
-
-  // Don't attempt to combine a point with the point below it if no point
-  // exists below it.
-  if (
-    !withinBounds(action.params.keepIndex) ||
-    !withinBounds(action.params.deleteIndex)
-  ) {
-    return state;
-  }
-
-  // Don't combine points with quoted points:
-  if (
-    isQuoted(action.params.keepIndex) ||
-    isQuoted(action.params.deleteIndex)
-  ) {
-    return state;
-  }
-
-  const pointIdToKeep =
-    currentMessage.shapes[action.params.shape][action.params.keepIndex];
-  const pointIdToDelete =
-    currentMessage.shapes[action.params.shape][action.params.deleteIndex];
-
-  const newContent =
-    action.params.keepIndex < action.params.deleteIndex
-      ? getPointById(pointIdToKeep, state).content +
-        getPointById(pointIdToDelete, state).content
-      : getPointById(pointIdToDelete, state).content +
-        getPointById(pointIdToKeep, state).content;
-  return produce(state, (draft) => {
-    delete draft.byId[pointIdToDelete];
-    getPointById(pointIdToKeep, draft).content = newContent;
-  });
-}
-
-function handleSplitIntoTwoPoints(
-  state: PointsState,
-  action: Action<_SplitIntoTwoPointsParams>
-): PointsState {
-  const topPoint = state.byId[action.params.pointId];
-  if (isReference(topPoint)) {
-    return state;
-  }
-
-  const topContent = topPoint.content.slice(0, action.params.sliceIndex);
-  const bottomContent = topPoint.content.slice(action.params.sliceIndex);
-
-  return produce(state, (draft) => {
-    getPointById(action.params.pointId, draft).content = topContent;
-    draft.byId[action.params.newPointId] = {
-      content: bottomContent,
-      _id: action.params.newPointId,
-      shape: topPoint.shape,
-      pointDate: new Date(),
-    };
-  });
-}
-
-function handleMessageCreate(
-  state: PointsState,
-  action: Action<_MessageCreateParams>
-): PointsState {
-  const { newReferencePoints } = action.params;
-
-  if (newReferencePoints === undefined) {
-    return state;
-  }
-
-  return produce(state, (draft) => {
-    newReferencePoints.forEach((point) => {
-      draft.byId[point._id] = point;
-    });
-  });
-}
-
-function handleMessageDelete(
-  state: PointsState,
-  action: Action<MessageDeleteParams>,
-  appState: AppState
-): PointsState {
-  const messageToDelete = appState.messages.byId[action.params.messageId];
-  let pointIdsToDelete = Object.values(messageToDelete.shapes).flat();
-
-  if (messageToDelete.focus !== undefined) {
-    pointIdsToDelete.push(messageToDelete.focus);
-  }
-
-  return produce(state, (draft) => {
-    pointIdsToDelete.forEach((id) => {
-      delete draft.byId[id];
-    });
-  });
-}
